// src/interfaces/interfaces.ts
export enum Lang {
  IT = "it",
  EN = "en",
  FR = "fr",
  ES = "es",
}
// ...
export enum TopicLevel {
  EASY,
  MEDIUM,
  HARD,
}